Split _base.py into 3 (step 2)
This needed to happen in 2 steps to allow for history preservation post split.

What was a single _base.py is now

- _crs.py (pyproj CRS wrapper)
- _geom.py (Shapely Geometry with CRS class and helpers)
- _geobox.py (GeoBox class and related utilities)

Note that there is a weak intentional cyclic dependency between `_crs.py` and
`_geom.py`. Mostly `_geom` needs CRS, that's a fundamental dependency. But CRS
also uses Geometry classes for some convenience methods, namely to report "valid
area" for a given CRS object.
