+++ odc/geo/__init__.py
@@ -4,30 +4,31 @@
 # SPDX-License-Identifier: Apache-2.0
 """ Geometric shapes and operations on them
 """
+# import order is important here
+#  _crs <-- _geom <-- _geobox <- other
+# isort: skip_file
 
 from ._version import __version__
 
-from ._base import (
+from ._crs import (
     CRS,
-    BoundingBox,
-    Coordinate,
-    CoordList,
     CRSError,
     CRSMismatchError,
-    GeoBox,
-    Geometry,
     MaybeCRS,
     SomeCRS,
-    assign_crs,
+    crs_units_per_degree,
+)
+
+from ._geom import (
+    BoundingBox,
+    CoordList,
+    Geometry,
     bbox_intersection,
     bbox_union,
     box,
     chop_along_antimeridian,
     clip_lon180,
     common_crs,
-    crs_units_per_degree,
-    geobox_intersection_conservative,
-    geobox_union_conservative,
     intersects,
     line,
     lonlat_bounds,
@@ -37,15 +38,25 @@ from ._base import (
     multipoint,
     multipolygon,
     point,
-    polygon,
     polygon_from_transform,
+    polygon,
     projected_lon,
-    scaled_down_geobox,
     sides,
     unary_intersection,
     unary_union,
 )
-from ._warp import rio_reproject, warp_affine
+
+from ._geobox import (
+    Coordinate,
+    GeoBox,
+    assign_crs,
+    geobox_intersection_conservative,
+    geobox_union_conservative,
+    scaled_down_geobox,
+)
+
+from ._gridspec import GridSpec
+
 from .tools import (
     affine_from_pts,
     apply_affine,
@@ -70,70 +81,70 @@ from .tools import (
     split_translation,
     w_,
 )
+from ._warp import rio_reproject, warp_affine
 
-from ._gridspec import GridSpec
 
 __all__ = [
-    "Coordinate",
+    "__version__",
+    "affine_from_pts",
+    "apply_affine",
+    "assign_crs",
+    "bbox_intersection",
+    "bbox_union",
     "BoundingBox",
+    "box",
+    "chop_along_antimeridian",
+    "clip_lon180",
+    "common_crs",
+    "compute_axis_overlap",
+    "compute_reproject_roi",
+    "Coordinate",
+    "CoordList",
+    "crs_units_per_degree",
+    "CRS",
     "CRSError",
     "CRSMismatchError",
-    "CRS",
-    "MaybeCRS",
-    "SomeCRS",
-    "CoordList",
-    "Geometry",
-    "GeoBox",
-    "GridSpec",
-    "assign_crs",
-    "common_crs",
-    "bbox_union",
-    "bbox_intersection",
-    "crs_units_per_degree",
-    "geobox_union_conservative",
+    "decompose_rws",
     "geobox_intersection_conservative",
+    "geobox_union_conservative",
+    "GeoBox",
+    "Geometry",
+    "get_scale_at_point",
+    "GridSpec",
     "intersects",
-    "point",
-    "multipoint",
-    "line",
-    "multiline",
-    "polygon",
-    "multipolygon",
-    "multigeom",
-    "box",
-    "sides",
-    "polygon_from_transform",
-    "unary_union",
-    "unary_intersection",
-    "lonlat_bounds",
-    "projected_lon",
-    "clip_lon180",
-    "chop_along_antimeridian",
-    "mid_longitude",
     "is_affine_st",
-    "apply_affine",
-    "compute_axis_overlap",
+    "line",
+    "lonlat_bounds",
+    "MaybeCRS",
+    "mid_longitude",
+    "multigeom",
+    "multiline",
+    "multipoint",
+    "multipolygon",
+    "native_pix_transform",
+    "point",
+    "polygon_from_transform",
+    "polygon",
+    "projected_lon",
+    "rio_reproject",
     "roi_boundary",
+    "roi_center",
+    "roi_from_points",
+    "roi_intersect",
     "roi_is_empty",
     "roi_is_full",
-    "roi_intersect",
-    "roi_shape",
     "roi_normalise",
-    "roi_from_points",
-    "roi_center",
     "roi_pad",
+    "roi_shape",
     "scaled_down_geobox",
-    "scaled_down_shape",
     "scaled_down_roi",
+    "scaled_down_shape",
     "scaled_up_roi",
-    "decompose_rws",
-    "affine_from_pts",
-    "get_scale_at_point",
-    "native_pix_transform",
-    "compute_reproject_roi",
+    "sides",
+    "SomeCRS",
     "split_translation",
-    "warp_affine",
-    "rio_reproject",
+    "unary_intersection",
+    "unary_union",
     "w_",
-    "__version__",
+    "warp_affine",
 ]
