+++ odc/geo/_geom.py
@@ -6,42 +6,25 @@ import array
 import functools
 import itertools
 import math
-import warnings
-from collections import OrderedDict, namedtuple
+from collections import namedtuple
 from collections.abc import Sequence
-from distutils.version import LooseVersion
-from typing import (
-    Any,
-    Callable,
-    Dict,
-    Hashable,
-    Iterable,
-    Iterator,
-    List,
-    Optional,
-    Tuple,
-    Union,
-)
+from typing import Any, Dict, Iterable, Iterator, List, Optional, Tuple, Union
 
-import cachetools
 import numpy
-import rasterio
-import xarray as xr
 from affine import Affine
-from pyproj.crs import CRS as _CRS
-from pyproj.enums import WktVersion
-from pyproj.exceptions import CRSError
-from pyproj.transformer import Transformer
 from shapely import geometry, ops
 from shapely.geometry import base
 
-from .math import is_almost_int
-from .tools import is_affine_st, roi_normalise, roi_shape
+from ._crs import (
+    CRS,
+    CRSMismatchError,
+    MaybeCRS,
+    SomeCRS,
+    _norm_crs,
+    _norm_crs_or_error,
+)
 
-Coordinate = namedtuple("Coordinate", ("values", "units", "resolution"))
 _BoundingBox = namedtuple("_BoundingBox", ("left", "bottom", "right", "top"))
-SomeCRS = Union[str, "CRS", _CRS, Dict[str, Any]]
-MaybeCRS = Optional[SomeCRS]
 CoordList = List[Tuple[float, float]]
 
 # pylint: disable=too-many-lines
@@ -124,250 +107,6 @@ class BoundingBox(_BoundingBox):
         :param p2: (x, y)
         """
         return BoundingBox.from_xy((p1[0], p2[0]), (p1[1], p2[1]))
-
-
-@cachetools.cached({})
-def _make_crs(crs_str: str) -> Tuple[_CRS, Optional[int]]:
-    crs = _CRS.from_user_input(crs_str)
-    return (crs, crs.to_epsg())
-
-
-def _make_crs_transform_key(from_crs, to_crs, always_xy):
-    return (id(from_crs), id(to_crs), always_xy)
-
-
-@cachetools.cached({}, key=_make_crs_transform_key)
-def _make_crs_transform(from_crs, to_crs, always_xy):
-    return Transformer.from_crs(from_crs, to_crs, always_xy=always_xy).transform
-
-
-def _guess_crs_str(crs_spec: Any) -> Optional[str]:
-    """
-    Returns a string representation of the crs spec.
-    Returns `None` if it does not understand the spec.
-    """
-    if isinstance(crs_spec, str):
-        return crs_spec
-    if isinstance(crs_spec, dict):
-        crs_spec = _CRS.from_dict(crs_spec)
-
-    if hasattr(crs_spec, "to_epsg"):
-        epsg = crs_spec.to_epsg()
-        if epsg is not None:
-            return f"EPSG:{crs_spec.to_epsg()}"
-    if hasattr(crs_spec, "to_wkt"):
-        return crs_spec.to_wkt()
-    return None
-
-
-class CRS:
-    """
-    Wrapper around `pyproj.CRS` for backwards compatibility.
-    """
-
-    DEFAULT_WKT_VERSION = (
-        WktVersion.WKT1_GDAL
-        if LooseVersion(rasterio.__gdal_version__) < LooseVersion("3.0.0")
-        else WktVersion.WKT2_2019
-    )
-
-    __slots__ = ("_crs", "_epsg", "_str")
-
-    def __init__(self, crs_str: Any):
-        """
-        :param crs_str: string representation of a CRS, often an EPSG code like 'EPSG:4326'
-        :raises: `pyproj.exceptions.CRSError`
-        """
-        crs_str = _guess_crs_str(crs_str)
-        if crs_str is None:
-            raise CRSError(
-                "Expect string or any object with `.to_epsg()` or `.to_wkt()` method"
-            )
-
-        _crs, _epsg = _make_crs(crs_str)
-
-        self._crs = _crs
-        self._epsg = _epsg
-        self._str = crs_str
-
-    def __getstate__(self):
-        return {"crs_str": self._str}
-
-    def __setstate__(self, state):
-        self.__init__(state["crs_str"])
-
-    def to_wkt(self, pretty: bool = False, version: Optional[WktVersion] = None) -> str:
-        """
-        WKT representation of the CRS
-        """
-        if version is None:
-            version = self.DEFAULT_WKT_VERSION
-
-        return self._crs.to_wkt(pretty=pretty, version=version)
-
-    @property
-    def wkt(self) -> str:
-        return self.to_wkt(version=WktVersion.WKT1_GDAL)
-
-    def to_epsg(self) -> Optional[int]:
-        """
-        EPSG Code of the CRS or None
-        """
-        return self._epsg
-
-    @property
-    def epsg(self) -> Optional[int]:
-        return self._epsg
-
-    @property
-    def semi_major_axis(self):
-        return self._crs.ellipsoid.semi_major_metre
-
-    @property
-    def semi_minor_axis(self):
-        return self._crs.ellipsoid.semi_minor_metre
-
-    @property
-    def inverse_flattening(self):
-        return self._crs.ellipsoid.inverse_flattening
-
-    @property
-    def geographic(self) -> bool:
-        return self._crs.is_geographic
-
-    @property
-    def projected(self) -> bool:
-        return self._crs.is_projected
-
-    @property
-    def dimensions(self) -> Tuple[str, str]:
-        """
-        List of dimension names of the CRS.
-        The ordering of the names is intended to reflect the `numpy` array axis order of the loaded raster.
-        """
-        if self.geographic:
-            return "latitude", "longitude"
-
-        if self.projected:
-            return "y", "x"
-
-        raise ValueError("Neither projected nor geographic")  # pragma: no cover
-
-    @property
-    def units(self) -> Tuple[str, str]:
-        """
-        List of dimension units of the CRS.
-        The ordering of the units is intended to reflect the `numpy` array axis order of the loaded raster.
-        """
-        if self.geographic:
-            return "degrees_north", "degrees_east"
-
-        if self.projected:
-            x, y = self._crs.axis_info
-            return x.unit_name, y.unit_name
-
-        raise ValueError("Neither projected nor geographic")  # pragma: no cover
-
-    def __str__(self) -> str:
-        return self._str
-
-    def __hash__(self) -> int:
-        return hash(self.to_wkt())
-
-    def __repr__(self) -> str:
-        return f"CRS('{self._str}')"
-
-    def __eq__(self, other) -> bool:
-        if not isinstance(other, CRS):
-            try:
-                other = CRS(other)
-            except Exception:  # pylint: disable=broad-except
-                return False
-
-        if self._crs is other._crs:
-            return True
-
-        if self.epsg is not None and other.epsg is not None:
-            return self.epsg == other.epsg
-
-        return self._crs == other._crs
-
-    def __ne__(self, other) -> bool:
-        return not self == other
-
-    @property
-    def proj(self) -> _CRS:
-        """Access proj.CRS object that this wraps"""
-        return self._crs
-
-    @property
-    def valid_region(self) -> Optional["Geometry"]:
-        """Return valid region of this CRS.
-
-        Bounding box in Lon/Lat as a 4 point Polygon in EPSG:4326.
-        None if not defined
-        """
-        aou = self._crs.area_of_use
-        if aou is None:
-            return None
-        return box(aou.west, aou.south, aou.east, aou.north, "EPSG:4326")
-
-    @property
-    def crs_str(self) -> str:
-        """DEPRECATED"""
-        warnings.warn(
-            "Please use `str(crs)` instead of `crs.crs_str`",
-            category=DeprecationWarning,
-        )
-        return self._str
-
-    def transformer_to_crs(
-        self, other: "CRS", always_xy=True
-    ) -> Callable[[Any, Any], Tuple[Any, Any]]:
-        """
-        Returns a function that maps x, y -> x', y' where x, y are coordinates in
-        this stored either as scalars or ndarray objects and x', y' are the same
-        points in the `other` CRS.
-        """
-
-        # pylint: disable=protected-access
-        transform = _make_crs_transform(self._crs, other._crs, always_xy=always_xy)
-
-        def result(x, y):
-            rx, ry = transform(x, y)  # pylint: disable=unpacking-non-sequence
-
-            if not isinstance(rx, numpy.ndarray) or not isinstance(ry, numpy.ndarray):
-                return (rx, ry)
-
-            missing = numpy.isnan(rx) | numpy.isnan(ry)
-            rx[missing] = numpy.nan
-            ry[missing] = numpy.nan
-            return (rx, ry)
-
-        return result
-
-
-class CRSMismatchError(ValueError):
-    """
-    Raised when geometry operation is attempted on geometries in different
-    coordinate references.
-    """
-
-
-def _norm_crs(crs: MaybeCRS) -> Optional[CRS]:
-    if isinstance(crs, CRS):
-        return crs
-    if crs is None:
-        return None
-    return CRS(crs)
-
-
-def _norm_crs_or_error(crs: MaybeCRS) -> CRS:
-    if isinstance(crs, CRS):
-        return crs
-    if crs is None:
-        raise ValueError("Expect valid CRS")
-    return CRS(crs)
 
 
 def wrap_shapely(method):
@@ -1008,352 +747,6 @@ def unary_intersection(geoms: Iterable[Geometry]) -> Geometry:
     return functools.reduce(Geometry.intersection, geoms)
 
 
-def _align_pix(left: float, right: float, res: float, off: float) -> Tuple[float, int]:
-    if res < 0:
-        res = -res
-        val = math.ceil((right - off) / res) * res + off
-        width = max(1, int(math.ceil((val - left - 0.1 * res) / res)))
-    else:
-        val = math.floor((left - off) / res) * res + off
-        width = max(1, int(math.ceil((right - val - 0.1 * res) / res)))
-    return val, width
-
-
-class GeoBox:
-    """
-    Defines the location and resolution of a rectangular grid of data,
-    including it's :py:class:`CRS`.
-
-    :param crs: Coordinate Reference System
-    :param affine: Affine transformation defining the location of the geobox
-    """
-
-    def __init__(self, width: int, height: int, affine: Affine, crs: MaybeCRS):
-        assert is_affine_st(
-            affine
-        ), "Only axis-aligned geoboxes are currently supported"
-        self.width = width
-        self.height = height
-        self.affine = affine
-        self.extent = polygon_from_transform(width, height, affine, crs=crs)
-
-    @classmethod
-    def from_geopolygon(
-        cls,
-        geopolygon: Geometry,
-        resolution: Tuple[float, float],
-        crs: MaybeCRS = None,
-        align: Optional[Tuple[float, float]] = None,
-    ) -> "GeoBox":
-        """
-        :param resolution: (y_resolution, x_resolution)
-        :param crs: CRS to use, if different from the geopolygon
-        :param align: Align geobox such that point 'align' lies on the pixel boundary.
-        """
-        align = align or (0.0, 0.0)
-        assert (
-            0.0 <= align[1] <= abs(resolution[1])
-        ), "X align must be in [0, abs(x_resolution)] range"
-        assert (
-            0.0 <= align[0] <= abs(resolution[0])
-        ), "Y align must be in [0, abs(y_resolution)] range"
-
-        if crs is None:
-            crs = geopolygon.crs
-        else:
-            geopolygon = geopolygon.to_crs(crs)
-
-        bounding_box = geopolygon.boundingbox
-        offx, width = _align_pix(
-            bounding_box.left, bounding_box.right, resolution[1], align[1]
-        )
-        offy, height = _align_pix(
-            bounding_box.bottom, bounding_box.top, resolution[0], align[0]
-        )
-        affine = Affine.translation(offx, offy) * Affine.scale(
-            resolution[1], resolution[0]
-        )
-        return GeoBox(crs=crs, affine=affine, width=width, height=height)
-
-    def buffered(self, ybuff, xbuff) -> "GeoBox":
-        """
-        Produce a tile buffered by ybuff, xbuff (in CRS units)
-        """
-        by, bx = (
-            _round_to_res(buf, res) for buf, res in zip((ybuff, xbuff), self.resolution)
-        )
-        affine = self.affine * Affine.translation(-bx, -by)
-
-        return GeoBox(
-            width=self.width + 2 * bx,
-            height=self.height + 2 * by,
-            affine=affine,
-            crs=self.crs,
-        )
-
-    def __getitem__(self, roi) -> "GeoBox":
-        if isinstance(roi, int):
-            roi = (slice(roi, roi + 1), slice(None, None))
-
-        if isinstance(roi, slice):
-            roi = (roi, slice(None, None))
-
-        if len(roi) > 2:
-            raise ValueError("Expect 2d slice")
-
-        if not all(s.step is None or s.step == 1 for s in roi):
-            raise NotImplementedError("scaling not implemented, yet")
-
-        roi = roi_normalise(roi, self.shape)
-        ty, tx = (s.start for s in roi)
-        h, w = roi_shape(roi)
-
-        affine = self.affine * Affine.translation(tx, ty)
-
-        return GeoBox(width=w, height=h, affine=affine, crs=self.crs)
-
-    def __or__(self, other) -> "GeoBox":
-        """A geobox that encompasses both self and other."""
-        return geobox_union_conservative([self, other])
-
-    def __and__(self, other) -> "GeoBox":
-        """A geobox that is contained in both self and other."""
-        return geobox_intersection_conservative([self, other])
-
-    def is_empty(self) -> bool:
-        return self.width == 0 or self.height == 0
-
-    def __bool__(self) -> bool:
-        return not self.is_empty()
-
-    def __hash__(self):
-        return hash((*self.shape, self.crs, self.affine))
-
-    @property
-    def transform(self) -> Affine:
-        return self.affine
-
-    @property
-    def shape(self) -> Tuple[int, int]:
-        return self.height, self.width
-
-    @property
-    def crs(self) -> Optional[CRS]:
-        return self.extent.crs
-
-    @property
-    def dimensions(self) -> Tuple[str, str]:
-        """
-        List of dimension names of the GeoBox
-        """
-        crs = self.crs
-        if crs is None:
-            return ("y", "x")
-        return crs.dimensions
-
-    @property
-    def resolution(self) -> Tuple[float, float]:
-        """
-        Resolution in Y,X dimensions
-        """
-        return self.affine.e, self.affine.a
-
-    @property
-    def alignment(self) -> Tuple[float, float]:
-        """
-        Alignment of pixel boundaries in Y,X dimensions
-        """
-        return self.affine.yoff % abs(self.affine.e), self.affine.xoff % abs(
-            self.affine.a
-        )
-
-    @property
-    def coordinates(self) -> Dict[str, Coordinate]:
-        """
-        dict of coordinate labels
-        """
-        yres, xres = self.resolution
-        yoff, xoff = self.affine.yoff, self.affine.xoff
-
-        xs = numpy.arange(self.width) * xres + (xoff + xres / 2)
-        ys = numpy.arange(self.height) * yres + (yoff + yres / 2)
-
-        units = self.crs.units if self.crs is not None else ("1", "1")
-
-        return OrderedDict(
-            (dim, Coordinate(labels, units, res))
-            for dim, labels, units, res in zip(
-                self.dimensions, (ys, xs), units, (yres, xres)
-            )
-        )
-
-    def xr_coords(
-        self, with_crs: Union[bool, str] = False
-    ) -> Dict[Hashable, xr.DataArray]:
-        """Dictionary of Coordinates in xarray format
-
-        :param with_crs: If True include netcdf/cf style CRS Coordinate
-        with default name 'spatial_ref', if with_crs is a string then treat
-        the string as a name of the coordinate.
-
-        Returns
-        =======
-
-        OrderedDict name:str -> xr.DataArray
-
-        where names are either `y,x` for projected or `latitude, longitude` for geographic.
-
-        """
-        spatial_ref = "spatial_ref"
-        if isinstance(with_crs, str):
-            spatial_ref = with_crs
-            with_crs = True
-
-        attrs = {}
-        coords = self.coordinates
-        crs = self.crs
-        if crs is not None:
-            attrs["crs"] = str(crs)
-
-        coords: Dict[Hashable, xr.DataArray] = {
-            n: _coord_to_xr(n, c, **attrs) for n, c in coords.items()
-        }
-
-        if with_crs and crs is not None:
-            coords[spatial_ref] = _mk_crs_coord(crs, spatial_ref)
-
-        return coords
-
-    @property
-    def geographic_extent(self) -> Geometry:
-        """GeoBox extent in EPSG:4326"""
-        if self.crs is None or self.crs.geographic:
-            return self.extent
-        return self.extent.to_crs(CRS("EPSG:4326"))
-
-    coords = coordinates
-    dims = dimensions
-
-    def __str__(self):
-        return f"GeoBox({self.geographic_extent})"
-
-    def __repr__(self):
-        return f"GeoBox({self.width}, {self.height}, {self.affine!r}, {self.crs})"
-
-    def __eq__(self, other):
-        if not isinstance(other, GeoBox):
-            return False
-
-        return (
-            self.shape == other.shape
-            and self.transform == other.transform
-            and self.crs == other.crs
-        )
-
-
-def bounding_box_in_pixel_domain(geobox: GeoBox, reference: GeoBox) -> BoundingBox:
-    """
-    Returns the bounding box of `geobox` with respect to the pixel grid
-    defined by `reference` when their coordinate grids are compatible,
-    that is, have the same CRS, same pixel size and orientation, and
-    are related by whole pixel translation,
-    otherwise raises `ValueError`.
-    """
-    tol = 1.0e-8
-
-    if reference.crs != geobox.crs:
-        raise ValueError("Cannot combine geoboxes in different CRSs")
-
-    a, b, c, d, e, f, *_ = ~reference.affine * geobox.affine
-
-    if not (
-        numpy.isclose(a, 1)
-        and numpy.isclose(b, 0)
-        and is_almost_int(c, tol)
-        and numpy.isclose(d, 0)
-        and numpy.isclose(e, 1)
-        and is_almost_int(f, tol)
-    ):
-        raise ValueError("Incompatible grids")
-
-    tx, ty = round(c), round(f)
-    return BoundingBox(tx, ty, tx + geobox.width, ty + geobox.height)
-
-
-def geobox_union_conservative(geoboxes: List[GeoBox]) -> GeoBox:
-    """Union of geoboxes. Fails whenever incompatible grids are encountered."""
-    if len(geoboxes) == 0:
-        raise ValueError("No geoboxes supplied")
-
-    reference, *_ = geoboxes
-
-    bbox = bbox_union(
-        bounding_box_in_pixel_domain(geobox, reference=reference) for geobox in geoboxes
-    )
-
-    affine = reference.affine * Affine.translation(*bbox[:2])
-
-    return GeoBox(
-        width=bbox.width, height=bbox.height, affine=affine, crs=reference.crs
-    )
-
-
-def geobox_intersection_conservative(geoboxes: List[GeoBox]) -> GeoBox:
-    """
-    Intersection of geoboxes. Fails whenever incompatible grids are encountered.
-    """
-    if len(geoboxes) == 0:
-        raise ValueError("No geoboxes supplied")
-
-    reference, *_ = geoboxes
-
-    bbox = bbox_intersection(
-        bounding_box_in_pixel_domain(geobox, reference=reference) for geobox in geoboxes
-    )
-
-    # standardise empty geobox representation
-    if bbox.left > bbox.right:
-        bbox = BoundingBox(
-            left=bbox.left, bottom=bbox.bottom, right=bbox.left, top=bbox.top
-        )
-    if bbox.bottom > bbox.top:
-        bbox = BoundingBox(
-            left=bbox.left, bottom=bbox.bottom, right=bbox.right, top=bbox.bottom
-        )
-
-    affine = reference.affine * Affine.translation(*bbox[:2])
-
-    return GeoBox(
-        width=bbox.width, height=bbox.height, affine=affine, crs=reference.crs
-    )
-
-
-def scaled_down_geobox(src_geobox: GeoBox, scaler: int) -> GeoBox:
-    """Given a source geobox and integer scaler compute geobox of a scaled down image.
-
-    Output geobox will be padded when shape is not a multiple of scaler.
-    Example: 5x4, scaler=2 -> 3x2
-
-    NOTE: here we assume that pixel coordinates are 0,0 at the top-left
-          corner of a top-left pixel.
-
-    """
-    assert scaler > 1
-
-    H, W = (X // scaler + (1 if X % scaler else 0) for X in src_geobox.shape)
-
-    # Since 0,0 is at the corner of a pixel, not center, there is no
-    # translation between pixel plane coords due to scaling
-    A = src_geobox.transform * Affine.scale(scaler, scaler)
-
-    return GeoBox(W, H, A, src_geobox.crs)
-
-
-def _round_to_res(value: float, res: float) -> int:
-    res = abs(res)
-    return int(math.ceil((value - 0.1 * res) / res))
-
-
 def intersects(a: Geometry, b: Geometry) -> bool:
     """Returns True if geometries intersect, else False"""
     return a.intersects(b) and not a.touches(b)
@@ -1391,63 +784,6 @@ def bbox_intersection(bbs: Iterable[BoundingBox]) -> BoundingBox:
         T = min(t, T)
 
     return BoundingBox(L, B, R, T)
-
-
-def _mk_crs_coord(crs: CRS, name: str = "spatial_ref") -> xr.DataArray:
-    # pylint: disable=protected-access
-
-    if crs.projected:
-        grid_mapping_name = crs._crs.to_cf().get("grid_mapping_name")
-        if grid_mapping_name is None:
-            grid_mapping_name = "??"
-        grid_mapping_name = grid_mapping_name.lower()
-    else:
-        grid_mapping_name = "latitude_longitude"
-
-    epsg = 0 if crs.epsg is None else crs.epsg
-
-    return xr.DataArray(
-        numpy.asarray(epsg, "int32"),
-        name=name,
-        dims=(),
-        attrs={"spatial_ref": crs.wkt, "grid_mapping_name": grid_mapping_name},
-    )
-
-
-def _coord_to_xr(name: str, c: Coordinate, **attrs) -> xr.DataArray:
-    """Construct xr.DataArray from named Coordinate object, this can then be used
-    to define coordinates for xr.Dataset|xr.DataArray
-    """
-    attrs = dict(units=c.units, resolution=c.resolution, **attrs)
-    return xr.DataArray(c.values, coords={name: c.values}, dims=(name,), attrs=attrs)
-
-
-def crs_units_per_degree(
-    crs: SomeCRS,
-    lon: Union[float, Tuple[float, float]],
-    lat: float = 0,
-    step: float = 0.1,
-) -> float:
-    """Compute number of CRS units per degree for a projected CRS at a given location
-    in lon/lat.
-
-    Location can be supplied as a tuple or as two arguments.
-
-    Returns
-    -------
-    A floating number S such that `S*degrees -> meters`
-    """
-    if isinstance(lon, tuple):
-        lon, lat = lon
-
-    lon2 = lon + step
-    if lon2 > 180:
-        lon2 = lon - step
-
-    ll = line([(lon, lat), (lon2, lat)], "EPSG:4326")
-    xy = ll.to_crs(crs, resolution=math.inf)
-
-    return xy.length / step
 
 
 def lonlat_bounds(
@@ -1492,52 +828,6 @@ def lonlat_bounds(
                 xx_range = xx_range_
 
     return BoundingBox.from_xy(xx_range, bbox.range_y)
-
-
-def assign_crs(
-    xx: Union[xr.DataArray, xr.Dataset],
-    crs: MaybeCRS = None,
-    crs_coord_name: str = "spatial_ref",
-) -> Union[xr.Dataset, xr.DataArray]:
-    """
-    Assign CRS for a non-georegistered array or dataset.
-
-    Returns a new object with CRS information populated.
-
-    Can also be called without ``crs`` argument on data that already has CRS
-    information but not in the format used by datacube, in this case CRS
-    metadata will be restructured into a shape used by datacube. This format
-    allows for better propagation of CRS information through various
-    computations.
-
-    .. code-block:: python
-
-        xx = datacube.utils.geometry.assign_crs(xr.open_rasterio("some-file.tif"))
-        print(xx.geobox)
-        print(xx.astype("float32").geobox)
-
-
-    :param xx:  Dataset or DataArray
-    :param crs: CRS to assign, if omitted try to guess from attributes
-    :param crs_coord_name: how to name crs corodinate (defaults to ``spatial_ref``)
-    """
-    if crs is None:
-        geobox = getattr(xx, "geobox", None)
-        if geobox is None:
-            raise ValueError("Failed to guess CRS for this object")
-        crs = geobox.crs
-
-    crs = _norm_crs_or_error(crs)
-    crs_coord = _mk_crs_coord(crs, name=crs_coord_name)
-    xx = xx.assign_coords({crs_coord.name: crs_coord})
-
-    xx.attrs.update(grid_mapping=crs_coord_name)
-
-    if isinstance(xx, xr.Dataset):
-        for band in xx.data_vars.values():
-            band.attrs.update(grid_mapping=crs_coord_name)
-
-    return xx
 
 
 def mid_longitude(geom: Geometry) -> float:
