+++ odc/geo/gbox.py
@@ -11,7 +11,8 @@ from typing import Dict, Iterable, Optional, Tuple
 
 from affine import Affine
 
-from ._base import BoundingBox, GeoBox, Geometry
+from ._geobox import GeoBox
+from ._geom import BoundingBox, Geometry
 from .math import clamp
 from .tools import align_up
